NXDRIVE-1542: Use new markers everywhere
- Added markers.linux_only
- Added markers.not_linux
- Added markers.not_mac

Also:
- Moved old_functionnal/test_windows.py::TestWindows::test_concurrent_file_access() to old_functional/test_concurrent_synchronization.py
- Moved old_functionnal/test_windows.py::TestWindows::test_local_replace() to old_functional/test_local_move_and_rename.py
- Moved old_functionnal/test_windows.py::TestWindows::test_registry_configuration() to functional/test_windows.py
- Moved old_functionnal/test_windows.py::test_registry_*() to unit/test_windows_registry.py
- Removed old_functional/test_windows.py
