+++ tests/markers.py
@@ -5,7 +5,7 @@ import os
 
 import pytest
 
-from nxdrive.constants import MAC, WINDOWS
+from nxdrive.constants import LINUX, MAC, WINDOWS
 
 
 # Skip tests based on data coming from Jenkins (not public)
@@ -13,11 +13,11 @@ jenkins_only = pytest.mark.skipif(
     "JENKINS_URL" not in os.environ, reason="Must be ran from Jenkins."
 )
 
-# Skip tests that must be run on macOS only
+# OS specific tests
+linux_only = pytest.mark.skipif(not LINUX, reason="GNU/Linux only.")
 mac_only = pytest.mark.skipif(not MAC, reason="macOS only.")
-
-# Skip tests that must be run on Windows only
 windows_only = pytest.mark.skipif(not WINDOWS, reason="Windows only.")
 
-# Skip tests that must __not__ be run on Windows
+not_linux = pytest.mark.skipif(LINUX)
+not_mac = pytest.mark.skipif(MAC)
 not_windows = pytest.mark.skipif(WINDOWS)
